Move encoding components to encoding submodule
