+++ relayer/crates/starknet-chain-components/src/components/mod.rs
@@ -1,4 +1,3 @@
-pub mod cairo_encoding;
 pub mod chain;
 pub mod cosmos_to_starknet;
-pub mod protobuf_encoding;
+pub mod encoding;
