+++ relayer/crates/starknet-chain-context/src/contexts/encoding/protobuf.rs
@@ -4,7 +4,7 @@ use hermes_encoding_components::traits::convert::CanConvert;
 use hermes_encoding_components::traits::encode_and_decode::CanEncodeAndDecode;
 use hermes_error::impls::ProvideHermesError;
 use hermes_protobuf_encoding_components::types::{Any, ViaProtobuf};
-use hermes_starknet_chain_components::components::protobuf_encoding::*;
+use hermes_starknet_chain_components::components::encoding::protobuf::*;
 use hermes_starknet_chain_components::types::client_state::{
     StarknetClientState, WasmStarknetClientState,
 };
